Split driver node into cpp and hpp files.
